Support single-pass paired-end trimming by adding -A/-G/-B parameters.
The -A/-G/-B parameters work just like their -a/-g/-b counterparts,
but specify adapters that are removed from the second read in a pair.

If one of the parameters is used, the read modification options such as
-q/-x/-y apply to *both* reads. For backwards compatibility, they do apply
only to the first read if neither of -A/-G/-B is used.

This is stil incomplete:
* --info-file is only written for first read
* -m/-M don’t work correctly.